new branch to reorganize main.py
